src/SPS/src/requirements.txt
python-dotenv==1.0.1
setuptools==69.1.1
python-snap7==1.3
flask-cors==4.0.0
requests==2.26.0
schedule==1.1.0